Convert to Spring project
