+++ Java/tollCalculator/src/main/java/com/acme/tollCalculator/Car.java
@@ -1,0 +1,9 @@
+package com.acme.tollCalculator;
+
+
+public class Car implements Vehicle {
+  @Override
+  public String getType() {
+    return "Car";
+  }
+}
