+++ Java/tollCalculator/src/main/java/com/acme/tollCalculator/Motorbike.java
@@ -1,0 +1,9 @@
+package com.acme.tollCalculator;
+
+
+public class Motorbike implements Vehicle {
+  @Override
+  public String getType() {
+    return "Motorbike";
+  }
+}
