+++ Java/tollCalculator/src/main/java/com/acme/tollCalculator/TollCalculatorApplication.java
@@ -1,0 +1,13 @@
+package com.acme.tollCalculator;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class TollCalculatorApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(TollCalculatorApplication.class, args);
+	}
+
+}
