+++ Java/tollCalculator/src/main/java/com/acme/tollCalculator/TollCalculatorController.java
@@ -1,0 +1,8 @@
+package com.acme.tollCalculator;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class TollCalculatorController {
+
+}
