+++ Java/tollCalculator/src/main/java/com/acme/tollCalculator/Vehicle.java
@@ -1,0 +1,7 @@
+package com.acme.tollCalculator;
+
+
+public interface Vehicle {
+
+  public String getType();
+}
