+++ Java/tollCalculator/src/test/java/com/acme/tollCalculator/TollCalculatorApplicationTests.java
@@ -1,0 +1,13 @@
+package com.acme.tollCalculator;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class TollCalculatorApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
